refactor: add mock suffix
