+++ tests/domain/mocks/queries/index.ts
@@ -1,2 +1,2 @@
-export * from './get-account-id-by-token'
+export * from './get-account-id-by-token-mock'
 export * from './verify-confirmation-code-middleware-mock'
